# backend/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "drmeet-api",
    "version": "1.0.0",
    "description": "A simple Appointment Booking System for small clinics, ready for frontend consumption"
  },
  "servers": [
    {
      "url": "https://drmeet-api.onrender.com/"
    },
    {
      "url": "http://127.0.0.1:3001"
    }
  ],
...
  "paths": {
...
    "/api/patients": {
      "get": {
        "summary": "Get all patients",
        "tags": [
          "Patients"
        ],
        "responses": {
          "200": {
            "description": "A list of all patients"
          },
          "500": {
            "description": "An error occurred while fetching patients"
          }
        }
      },
      "post": {
        "summary": "Create a new patient",
        "tags": [
          "Patients"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "firstName",
                  "lastName",
                  "email"
                ],
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "A new patient created"
          },
          "400": {
            "description": "Failed to create patient"
          },
          "500": {
            "description": "An error occurred while creating the patient"
          }
        }
      }
    },
    "/api/patients/{id}": {
      "get": {
        "summary": "Get a patient by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the patient"
          }
        ],
        "tags": [
          "Patients"
        ],
        "responses": {
          "200": {
            "description": "The patient with the specified ID"
          },
          "400": {
            "description": "Invalid patient ID format"
          },
          "404": {
            "description": "Patient not found"
          },
          "500": {
            "description": "An error occurred while fetching the patient"
          }
        }
      },
      "put": {
        "summary": "Update a patient by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the patient"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "firstName": {
                    "type": "string"
                  },
                  "lastName": {
                    "type": "string"
                  },
                  "email": {
                    "type": "string"
                  },
                  "phone": {
                    "type": "string"
                  },
                  "address": {
                    "type": "string"
                  },
                  "dateOfBirth": {
                    "type": "string",
                    "format": "date"
                  }
                }
              }
            }
          }
        },
        "tags": [
          "Patients"
        ],
        "responses": {
          "200": {
            "description": "Patient updated"
          },
          "400": {
            "description": "Invalid patient ID format"
          },
          "404": {
            "description": "Patient not found"
          },
          "500": {
            "description": "An error occurred while updating the patient"
          }
        }
      },
      "delete": {
        "summary": "Deletes a patient by ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "The unique ID of the patient"
          }
        ],
        "tags": [
          "Patients"
        ],
        "responses": {
          "200": {
            "description": "Patient was deleted successfully"
          },
          "400": {
            "description": "Invalid patient ID format"
          },
          "404": {
            "description": "Patient not found"
          },
          "500": {
            "description": "An error occurred while deleting the patient"
          }
        }
      }
    }
  }
}